% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlearning_CVAE.R
\name{cvae_qlearning}
\alias{cvae_qlearning}
\title{CVAE Q-learning Method}
\usage{
cvae_qlearning(X, A, Y, latent_dim = 3, epochs = 100)
}
\arguments{
\item{X}{Covariate matrix}

\item{A}{Treatment matrix or categorical/one-hot}

\item{Y}{Outcome}

\item{latent_dim}{Dimension of latent variable Z}

\item{epochs}{Number of training epochs}
}
\value{
A list containing encoder, decoder, Q model, and treatment decoder
}
\description{
CVAE Q-learning Method
}
