% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtr_gof.R
\name{dtr_gof}
\alias{dtr_gof}
\title{Goodness-of-Fit for dtr() Output}
\usage{
dtr_gof(dtr_result, outcome = "y")
}
\arguments{
\item{dtr_result}{Output from `dtr()`.}

\item{outcome}{Name of the outcome variable (default = "y").}
}
\value{
A list of goodness-of-fit metrics.
}
\description{
Computes goodness-of-fit metrics: RMSE, MAE, and R-squared if applicable.
}
