% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\docType{data}
\name{f1}
\alias{f1}
\title{Example Q-learning Formula}
\format{
An object of class \code{formula}.
}
\usage{
data(f1)
}
\description{
A formula object used to specify an interaction model for Q-learning.
This includes main effects and interactions between treatment `a` and covariates `x1` through `x10`.
}
\keyword{datasets}
