% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dtr_forest.R
\name{plot_dtr_forest}
\alias{plot_dtr_forest}
\title{Forest Plot for dtr() Output with 95% Interval and Color}
\usage{
plot_dtr_forest(dtr_result, method = c("BayesianQ", "Lasso", "ElasticNet"))
}
\arguments{
\item{dtr_result}{The output list from `dtr()` function.}

\item{method}{One of "BayesianQ", "Lasso", "ElasticNet".}
}
\value{
A ggplot2 object.
}
\description{
Generates a colorful forest plot for Bayesian posterior means or non-zero coefficients from Lasso/ElasticNet.
}
