% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlearning_baysian.R
\name{qlearning_bayes}
\alias{qlearning_bayes}
\title{Bayesian Weighted Q-learning (Skeleton)}
\usage{
qlearning_bayes(data, formula = f1, testdat)
}
\arguments{
\item{X}{Covariates matrix.}

\item{A}{Treatment vector.}

\item{Y}{Outcome vector.}

\item{prior_weight}{Prior weight parameter.}
}
\value{
Placeholder for Bayesian model output.
}
\description{
Bayesian Weighted Q-learning (Skeleton)
}
