% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlearning_dyn.R
\name{qlearning_dyn}
\alias{qlearning_dyn}
\title{Wrapper for DynTxRegime::qLearn}
\usage{
qlearning_dyn(data, formula = f1, testdat)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Additional arguments to `DynTxRegime::qLearn()`.}
}
\value{
Output from `DynTxRegime::qLearn`.
}
\description{
Wrapper for DynTxRegime::qLearn
}
