% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlearning_gof_plot.R
\name{qlearning_gof_plot}
\alias{qlearning_gof_plot}
\title{Goodness-of-Fit Plot for Q-learning Models}
\usage{
qlearning_gof_plot(dtr_result, outcome = "y")
}
\arguments{
\item{dtr_result}{A result list returned from the `dtr()` function.}

\item{outcome}{A character string indicating the name of the outcome variable. Default is `"y"`.}
}
\value{
A ggplot object showing observed vs. predicted values, including a 45-degree reference line.
}
\description{
This function generates a scatter plot comparing observed vs. predicted outcomes
from a fitted Q-learning model object (either `lm` or `cv.glmnet`) as returned by `dtr()`.
}
\examples{
data <- data.frame(x1 = rnorm(100), x2 = rnorm(100), a = factor(sample(1:2, 100, TRUE)))
data$y <- with(data, 2 * as.numeric(a) + 1.5 * x1 - 0.5 * x2 + rnorm(100))
fit <- dtr(data = data, formula = y ~ x1 + x2 + a, method = "Qlearning", treatment = "a", outcome = "y")
p <- qlearning_gof_plot(fit)
print(p)
}
