% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlearning_lasso.R
\name{qlearning_lasso}
\alias{qlearning_lasso}
\title{Q-learning with Lasso Regularization}
\usage{
qlearning_lasso(data, formula = f1)
}
\arguments{
\item{X}{A matrix of covariates.}

\item{A}{A binary treatment vector (e.g., -1, 1).}

\item{Y}{Outcome vector.}

\item{lambda}{Optional lambda value; if NULL, cross-validation will be used.}
}
\value{
A fitted `cv.glmnet` object.
}
\description{
Fits a Q-learning model using a lasso penalty via the `glmnet` package.
}
