% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Simulated Q-learning Example Dataset}
\format{
A data frame with 200 observations and 7 variables:
\describe{
  \item{Y}{Continuous outcome}
  \item{A}{Binary treatment indicator (-1 or 1)}
  \item{X1--X5}{Covariates}
}
}
\usage{
data(sample_data)
}
\description{
A simulated dataset for demonstrating Q-learning methods.
}
\keyword{datasets}
